+++ vk-diff/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "aptos-keyless-common"
-description = "Aptos Keyless circuit tests"
+name = "aptos-keyless-vk-diff"
+description = "Aptos Keyless VK diff tool"
 version = "0.1.0"
 
 # Workspace inherited keys
@@ -17,10 +17,18 @@ anyhow = { workspace = true }
 aptos-types = { workspace = true }
 ark-bn254 = { workspace = true }
 ark-ff = { workspace = true }
+ark-serialize = { workspace = true }
 base64 = { workspace = true }
+clap = { workspace = true }
 hex = { workspace = true }
 num-bigint = { workspace = true }
+num-traits = {  workspace = true }
+once_cell = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
+strum = { workspace =  true }
+strum_macros = { workspace = true }
 tempfile = { workspace = true }
 tracing = { workspace = true }
+ureq = { workspace = true }
+url = { workspace = true }
